refactor: organize SQL validation modules
